Fix: Variable Definition

+++ esphome/components/seeed_mr60fda2/seeed_mr60fda2.h
@@ -85,7 +85,7 @@ class MR60FDA2Component : public Component,
   uint16_t current_frame_type_;
   uint32_t current_install_height_int_;
   uint32_t current_height_threshold_int_;
-  uint32_t current_sensitivity_;
+  uint8_t current_sensitivity_;
   uint8_t select_index_;
 
   bool validate_checksum_(const uint8_t *data, size_t len, uint8_t expected_checksum);
